ctype sign fixes (Tom Miller)

+++ src/cmd/grap/input.c
@@ -277,9 +277,9 @@ nextchar(void)
 				ERROR "argfp underflow" FATAL;
 			popsrc();
 			goto loop;
-		} else if (c == '$' && isdigit(*srcp->sp)) {	/* $3 */
+		} else if (c == '$' && isdigit((uchar)*srcp->sp)) {	/* $3 */
 			int n = 0;
-			while (isdigit(*srcp->sp))
+			while (isdigit((uchar)*srcp->sp))
 				n = 10 * n + *srcp->sp++ - '0';
 			if (n > 0 && n <= MAXARGS)
 				pushsrc(String, argfp->argstk[n-1]);
@@ -456,11 +456,11 @@ void eprint(void)	/* try to print context around error */
 	for (q=ep-1; q>=p && *q!=' ' && *q!='\t' && *q!='\n'; q--)
 		;
 	for (; p < q; p++)
-		if (isprint(*p))
+		if (isprint((uchar)*p))
 			putc(*p, stderr);
 	fprintf(stderr, " >>> ");
 	for (; p < q; p++)
-		if (isprint(*p))
+		if (isprint((uchar)*p))
 			putc(*p, stderr);
 	fprintf(stderr, " <<< ");
 	while (pb >= pbuf)
